Simplify ofo-queue
We cannot increase the size of the sk_buff. Thus, the pointers and
fields in sk_buff have to be removed.

Ofo-queue handling follows now this algorithm in try_shortcut():

First, we check the shortcut of the receiving subflow.
Then, we check the shortcut of the other subflows, and look for the skb
in the ofo-queue who is closest to the point where we want to insert the
skb. That way we reduce the lookup-time as we start iterating over the
ofo-queue from this point on.

This also removes the allocation of the mptcp_node in the ofo-algo and
thus reduces the number of cache-misses.

Signed-off-by: Christoph Paasch <christoph.paasch@uclouvain.be>
Reviewed-by: Octavian Purdila <octavian.purdila@intel.com>
Signed-off-by: Christoph Paasch <christoph.paasch@uclouvain.be>