Fix: Always handle subflow-fin after processing the rcv-queue
If the data-fin got dropped and then a subflow-fin is sent. And at the
retransmission of the data-fin the subflow-fin comes from the
subflow-ofo-queue, the socket will get closed in tcp_fin.

The fix is to always NOT close the socket in tcp_fin. It will get closed
in mptcp_data_ready, as the callback sk_data_ready is always called if
the socket is an MPTCP-subflow.

Signed-off-by: Christoph Paasch <christoph.paasch@uclouvain.be>
Signed-off-by: Gregory Detal <gregory.detal@uclouvain.be>