Fix: Don't reinject if meta is closed
In that case we passed already via tcp_v4_destroy_sock and thus purged
the reinject-queue. It won't get purged again, and we would have a
memory-leak if we add a segment to the reinject-queue.

So, don't reinject if the meta is in TCP_CLOSE.

Further, don't reinject empty subflow fins with a data-fin. They are
reinjected by the code at the end of mptcp_reinject_data, without the
tcp_flag on TCPHDR_FIN.

Signed-off-by: Christoph Paasch <christoph.paasch@uclouvain.be>

Conflicts:
	net/mptcp/mptcp_output.c

Signed-off-by: Gregory Detal <gregory.detal@uclouvain.be>