Fix: pskb_copy reinjected skb's if they are cloned
In case of address-removal the send-queue is purged and everything is
put in the reinject-queue. However, these segments may still be a clone
from tcp_transmit_skb, waiting in the qdisc.

When sending a reinjected segment we have to check for skb_cloned() and
do a pskb_copy in that case.

Signed-off-by: Christoph Paasch <christoph.paasch@uclouvain.be>
Signed-off-by: Gregory Detal <gregory.detal@uclouvain.be>