Fix: Purge sub-rcv-queue only when the full mapping is present.
Because, we may have a fallback otherwise:

We receive the first segment who holds the mapping (subsequent ones
don't). As this dss has already been received, the packet will be purged
from the rcv-queue. As the subsequent packets come in, a fallback will
be initiated, as no dss-option is present and the mapping is unknown.

Signed-off-by: Christoph Paasch <christoph.paasch@uclouvain.be>