Fix: Rebuild the mapping on the subflow when fragmenting or trimming
If we loose a mapping due to fragmenting or trimming, the reinjection
cannot work. This will then require the meta-level retransmission timer
to trigger the recovery. This is very slow and may even lead to stalls.

We now reconstruct the mapping when fragmenting or trimming. This allows
us to quickly recover from the loss. In __mptcp_reinject_data, we have
to look for the data from the meta-send-queue in order to reinject it
properly.

Signed-off-by: Christoph Paasch <christoph.paasch@uclouvain.be>
Signed-off-by: Gregory Detal <gregory.detal@uclouvain.be>