First attempt at fixing infinite mapping
When accessing baidu.com, it replies with the MP_CAPABLE option (because
it seems to just echo back any unknown option to the sender). However,
further baidu.com off course does not support MPTCP and will reply to
our data-seq without a data-ack. This is an easy scenario to trigger a
fallback to infinite through Section 3.6 of the RFC.

This is a first attempt to contacting baidu.com work.

Signed-off-by: Christoph Paasch <christoph.paasch@uclouvain.be>
Signed-off-by: Gregory Detal <gregory.detal@uclouvain.be>