Fix: Store the original values of snd/rcv-buffers and use them
If we don't do this, the server-side new subflows have a bloated up
snd/rcv-buffer (as it's a clone from the meta). This will then end up in
strange values for rcv_wnd of the new subflow, and incorrect setting of
the subflow-level rcv_ssthresh. As the autotuning during slow-start is
handled by rcv_ssthresh, this incorrect setting prevents correct
autotuning.

This patch makes a first step, by storing the original values of
snd/rcv-buffer in the mpcb.
New subflows on the server-side will thus get correct buffer-sizes. This
will allow in a follow-up patch to correctly increase the
snd/rcv-buffers (and meta-level rcv_ssthresh) upon establishment of
new subflows and thus allow a proper autotuning during slowstart.

Signed-off-by: Christoph Paasch <christoph.paasch@uclouvain.be>